backend/requirements.txt
flask
flask-cors
pymongo
requests
twilio
certifi
scikit-learn
python-dotenv
dnspython